requirements.txt
fastapi
uvicorn
transformers
torch
google-generativeai